+++ src/parser/api/parse.test.ts
@@ -38,8 +38,11 @@ describe('Unit: parser/api/parse', () => {
         [...parse('a\n\\\nb').children].map(el => el.outerHTML),
         ['<p>a</p>', '<p>b</p>']);
       assert.deepStrictEqual(
-        [...parse('~~~\ninvalid\n\ncaption\n~~~').children].map(el => el.outerHTML),
+        [...parse('~~~\ninvalid\n~~~').children].map(el => el.outerHTML),
         ['<p class="invalid" data-invalid-type="syntax">Invalid syntax: Extension: Invalid extension name, attribute, or content.</p>']);
+      assert.deepStrictEqual(
+        [...parse('~~~\ninvalid\n\ncaption\n~~~').children].map(el => el.outerHTML),
+        ['<p class="invalid" data-invalid-type="syntax">Invalid syntax: Extension: Invalid extension name, attribute, or content.</p>', '<p>caption<span class="linebreak"> </span>~~~</p>']);
     });
 
     it('normalize', () => {