+++ src/parser/block/extension.test.ts
@@ -1,22 +1,23 @@
 ﻿import { extension } from './extension';
 import { some } from '../../combinator';
-import { inspect } from '../../debug.test';
 
 describe('Unit: parser/block/extension', () => {
   describe('extension', () => {
     const parser = some(extension);
 
     it('invalid', () => {
-      assert.deepStrictEqual(inspect(parser('')), undefined);
-      assert.deepStrictEqual(inspect(parser('\n')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~~')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~~\n')), undefined);
-      assert.deepStrictEqual(inspect(parser('~~~\na~~~')), undefined);
-      assert.deepStrictEqual(inspect(parser(' ~~~\n~~~')), undefined);
+      assert(!parser(''));
+      assert(!parser('\n'));
+      assert(!parser(' ~~~\n~~~'));
     });
 
     it('valid', () => {
+      assert(parser('~~~'));
+      assert(parser('~~~\n'));
+      assert(parser('~~~\n\n'));
+      assert(parser('~~~\na~~~'));
       assert(parser('~~~\n~~~'));
+      assert(parser('~~~\na\n~~~'));
     });
 
   });